revert: remove latency predictor sidecar injection (#5678)
Reverts the latency predictor sidecar injection feature added in #5678.
The auto-revert failed due to conflicts from subsequent changes to the
same files.

The hasPluginType helper introduced in #5678 is retained in
tokenizer.go as it is used by the tokenizer sidecar logic added in

#5712.
