+++ config/llmisvcconfig/kustomization.yaml
@@ -10,7 +10,6 @@ resources:
   - config-llm-prefill-worker-data-parallel.yaml
   - config-llm-router-route.yaml
   - config-llm-scheduler.yaml
-  - config-llm-scheduler-latency-predictor.yaml
   - config-llm-template.yaml
   - config-llm-tokenizer.yaml
   - config-llm-tracing.yaml
